+++ src/pages/CV/components/Rating/styles.js
